Refresh wording to match latest version of charter template
Charter template is visible at:
https://w3c.github.io/charter-drafts/charter-template.html

Main change is the mention of the Code of Ethics and Professional Conduct.

Some minor changes to adjust wording here and there, notably in the Coordination
section. The updated text does not use the exact same wording as in the template
because the template is focused on groups producing normative specifications,
and the Interest Group will not develop such specifications.