Get back to boilerplate text for Coordination section
Previous text was an attempt to adjust the boilerplate text to account for the
fact that the IG won't publish any document in CR and, most of the time, will
not even publish an IG Note.

However, that text was clumsy, as pointed out in:
https://github.com/w3c/media-and-entertainment/issues/52

This update brings the text back to what it is in the charter template. The
bits on CR do not really apply to the IG, but then the group could perhaps one
day decide to publish non normative best practices that could go to Rec, so it
seems a good idea to keep the text around.