Add proposed charter for a performance task force
This proposed charter merges proposed texts in #86 and #88, following a similar
structure to previously created task forces (e.g.
https://www.w3.org/2011/webtv/wiki/Main_Page/Media_Timed_Events_TF).

In particular:
- It adds a "mission" section that lists the tasks that the task force would be
responsible for.
- It adds a "success criteria" section that lists deliverables.

The text also include an "Out-of-scope" section to note that the TF will *not*
create a benchmark of existing consumer products.

Note: the text uses the term "consumer products" instead of "TV" or "embedded
device". The same term is used in the Web Media API document:
https://w3c.github.io/webmediaapi/#introduction

The creation of a toolkit to measure performance of web UIs on consumer products
could be further refined (also see comment from @jpiesing in
https://github.com/w3c/media-and-entertainment/pull/88#issuecomment-1195108292).
I'm not clear what the group is willing to do on that front so left that as-is.