Refresh boilerplate text
This refreshes the boilerplate according to the latest charter template at:
https://w3c.github.io/charter-drafts/charter-template.html

Main updates are to mention the Ethical Web Principles and Privacy Principles
on top of the Web Platform Design Principles, and to refresh the URLs to the
Process and other policy documents.

The charter uses "deliverables" in prose instead of "specifications" since it
will not publish specifications on the Recommendation track. This update also
replaces a couple of occurrences where "specifications" was still used.