Create draft 2025 version of charter (#111)
* Create 2025 version out of approved 2023 charter

* Refresh boilerplate text

This refreshes the boilerplate according to the latest charter template at:
https://w3c.github.io/charter-drafts/charter-template.html

Main updates are to mention the Ethical Web Principles and Privacy Principles
on top of the Web Platform Design Principles, and to refresh the URLs to the
Process and other policy documents.

The charter uses "deliverables" in prose instead of "specifications" since it
will not publish specifications on the Recommendation track. This update also
replaces a couple of occurrences where "specifications" was still used.

* Add Background and Motivation section

This adds a background and motivation section at the beginning of the charter
as envisioned in the charter template. The end of the section is an updated
version of the text that appeared in the "Charter history" section.

* Refresh coordination list

This refreshes the list in the Coordination section, adding SVTA (which now
hosts DASH-IF), TTA and 3GPP.

Remove closed or soon to be closed groups.
Add DataCue under WICG.

---------

Co-authored-by: Chris Needham <chris@chrisneedham.com>