[charter] Expand on device independence
Feedback received during the Advisory Committee review of the charter suggests
calling out the need to ensure timely updates to client runtimes to avoid
fragmentation of the web platform.

The success criteria section already highlighted the Ethical Web principles,
which defines the web as multi-browser, multi-OS, and multi-device, as
a guiding document, and listed device independence as one of the issues to
consider in all discussions and outcomes for the Interest Group.

This update makes the criterion more explicit, by expanding the mention of
device independence to note that it includes ensuring timely updates to
client runtimes.